% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Bacon.R, R/BaconClam.R
\name{Bacon}
\alias{Bacon}
\title{Run Bacon.}
\usage{
Bacon(core = "MSB2K", thick = 5, prob = 0.95, d.min = NA, d.max = NA,
  d.by = 1, unit = "cm", maxcalc = 500, depths.file = FALSE,
  acc.shape = 1.5, acc.mean = 20, mem.strength = 4, mem.mean = 0.7,
  hiatus.depths = NA, hiatus.shape = 1, hiatus.mean = 1000,
  after = 1e-04, cc = 3, cc1 = "IntCal13", cc2 = "Marine13",
  cc3 = "SHCal13", cc4 = "ConstCal", postbomb = 0, d.R = 0, d.STD = 0,
  t.a = 3, t.b = 4, normal = FALSE, suggest = TRUE, reswarn = c(10,
  200), remember = TRUE, cleanup = TRUE, ask = TRUE, run = TRUE,
  defaults = "default_settings.txt", sep = ",", dec = ".", runname = "",
  slump = NA, BCAD = FALSE, ssize = 2000, rounded = 1, th0 = c(),
  burnin = min(200, ssize), MinYr = c(), MaxYr = c(), find.round = 4,
  bins = c(), cutoff = 0.001, plot.pdf = TRUE, rotate.axes = FALSE,
  rev.yr = FALSE, rev.d = FALSE, yr.min = c(), yr.max = c(),
  normalise.dists = TRUE, plot.title = core, title.location = "topleft",
  d.lab = "Depth", yr.lab = c(), d.res = 200, yr.res = 200,
  date.res = 100, grey.res = 100, width = 15, dark = 1,
  greyscale = function(x) grey(1 - x), C14.col = rgb(0, 0, 1, 0.35),
  C14.border = rgb(0, 0, 1, 0.5), cal.col = rgb(0, 0.5, 0.5, 0.35),
  cal.border = rgb(0, 0.5, 0.5, 0.5), range.col = grey(0.5),
  range.lty = "12", hiatus.col = grey(0.5), hiatus.lty = "12",
  wm.col = "red", wm.lty = "12", med.col = NA, med.lty = "12",
  mar = c(3, 3, 1, 1), mgp = c(1.5, 0.7, 0), bty = "l",
  saveDATA = FALSE)

Bacon(core = "MSB2K", thick = 5, prob = 0.95, d.min = NA, d.max = NA,
  d.by = 1, unit = "cm", maxcalc = 500, depths.file = FALSE,
  acc.shape = 1.5, acc.mean = 20, mem.strength = 4, mem.mean = 0.7,
  hiatus.depths = NA, hiatus.shape = 1, hiatus.mean = 1000,
  after = 1e-04, cc = 3, cc1 = "IntCal13", cc2 = "Marine13",
  cc3 = "SHCal13", cc4 = "ConstCal", postbomb = 0, d.R = 0, d.STD = 0,
  t.a = 3, t.b = 4, normal = FALSE, suggest = TRUE, reswarn = c(10,
  200), remember = TRUE, cleanup = TRUE, ask = TRUE, run = TRUE,
  defaults = "default_settings.txt", sep = ",", dec = ".", runname = "",
  slump = NA, BCAD = FALSE, ssize = 2000, rounded = 1, th0 = c(),
  burnin = min(200, ssize), MinYr = c(), MaxYr = c(), find.round = 4,
  bins = c(), cutoff = 0.001, plot.pdf = TRUE, rotate.axes = FALSE,
  rev.yr = FALSE, rev.d = FALSE, yr.min = c(), yr.max = c(),
  normalise.dists = TRUE, plot.title = core, title.location = "topleft",
  d.lab = "Depth", yr.lab = c(), d.res = 200, yr.res = 200,
  date.res = 100, grey.res = 100, width = 15, dark = 1,
  greyscale = function(x) grey(1 - x), C14.col = rgb(0, 0, 1, 0.35),
  C14.border = rgb(0, 0, 1, 0.5), cal.col = rgb(0, 0.5, 0.5, 0.35),
  cal.border = rgb(0, 0.5, 0.5, 0.5), range.col = grey(0.5),
  range.lty = "12", hiatus.col = grey(0.5), hiatus.lty = "12",
  wm.col = "red", wm.lty = "12", med.col = NA, med.lty = "12",
  mar = c(3, 3, 1, 1), mgp = c(1.5, 0.7, 0), bty = "l",
  saveDATA = FALSE)
}
\arguments{
\item{allparameters}{Usual parameters from the Bacon script. Check the manual.}

\item{saveData}{Boolean to keep Bacon outputs in the environment. Default: FALSE.}

\item{allparameters}{Usual parameters from the Bacon script. Check the manual.}

\item{saveData}{Boolean to keep Bacon outputs in the environment. Default: FALSE.}
}
\description{
Run Bacon.

Run Bacon.
}
\examples{
Bacon()
Bacon()
}
\keyword{Age-depth}
\keyword{Chronology}
\keyword{model}

