% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Bacon.R, R/BaconClam.R
\name{Clam}
\alias{Clam}
\title{Run Clam.}
\usage{
Clam(name = "Example", type = 1, smooth = c(), prob = 0.95,
  its = 1000, wghts = 1, cc = 3, cc1 = "IntCal13.14C",
  cc2 = "Marine13.14C", cc3 = "SHCal13.14C", cc4 = "mixed.14C",
  cc5 = "gluedHemispheres.14C", postbomb = FALSE,
  pb1 = "postbomb_NH1.14C", pb2 = "postbomb_NH2.14C",
  pb3 = "postbomb_NH3.14C", pb4 = "postbomb_SH1-2.14C",
  pb5 = "postbomb_SH3.14C", outliers = c(), ignore = c(),
  youngest = c(), extradates = c(), slump = c(), est = 1,
  calibt = FALSE, mixed.effect = FALSE, dmin = c(), dmax = c(),
  every = 1, yrmin = c(), yrmax = c(), yrsteps = 1, pbsteps = 0.01,
  hpdsteps = 1, BCAD = FALSE, decimals = 0, accrate = 0,
  ageofdepth = c(), depth = "cm", depthseq = c(), depths.file = FALSE,
  thickness = 1, hiatus = c(), remove.reverse = 0.5, times = 5,
  sep = ",", ext = ".csv", runname = c(), storedat = FALSE,
  threshold = 1e-06, proxies = FALSE, revaxes = FALSE, revd = TRUE,
  revyr = TRUE, calhght = 0.3, maxhght = 0.01, mirror = TRUE,
  plotrange = TRUE, bty = "l", mar = c(3.5, 3, 2, 1), mgp = c(2, 1, 0),
  plotpdf = TRUE, plotpng = TRUE, greyscale = c(), yrlab = c(),
  dlab = c(), calcol = rgb(0, 0.5, 0.5, 0.5), C14col = rgb(0, 0, 1, 0.5),
  outcol = "red", outlsize = 1, bestcol = "black", rangecol = rgb(0, 0,
  0, 0.3), slumpcol = grey(0.75), plotname = TRUE, ash = FALSE)

Clam(name = "Example", type = 1, smooth = c(), prob = 0.95,
  its = 1000, wghts = 1, cc = 3, cc1 = "IntCal13.14C",
  cc2 = "Marine13.14C", cc3 = "SHCal13.14C", cc4 = "mixed.14C",
  cc5 = "gluedHemispheres.14C", postbomb = FALSE,
  pb1 = "postbomb_NH1.14C", pb2 = "postbomb_NH2.14C",
  pb3 = "postbomb_NH3.14C", pb4 = "postbomb_SH1-2.14C",
  pb5 = "postbomb_SH3.14C", outliers = c(), ignore = c(),
  youngest = c(), extradates = c(), slump = c(), est = 1,
  calibt = FALSE, mixed.effect = FALSE, dmin = c(), dmax = c(),
  every = 1, yrmin = c(), yrmax = c(), yrsteps = 1, pbsteps = 0.01,
  hpdsteps = 1, BCAD = FALSE, decimals = 0, accrate = 0,
  ageofdepth = c(), depth = "cm", depthseq = c(), depths.file = FALSE,
  thickness = 1, hiatus = c(), remove.reverse = 0.5, times = 5,
  sep = ",", ext = ".csv", runname = c(), storedat = FALSE,
  threshold = 1e-06, proxies = FALSE, revaxes = FALSE, revd = TRUE,
  revyr = TRUE, calhght = 0.3, maxhght = 0.01, mirror = TRUE,
  plotrange = TRUE, bty = "l", mar = c(3.5, 3, 2, 1), mgp = c(2, 1, 0),
  plotpdf = TRUE, plotpng = TRUE, greyscale = c(), yrlab = c(),
  dlab = c(), calcol = rgb(0, 0.5, 0.5, 0.5), C14col = rgb(0, 0, 1, 0.5),
  outcol = "red", outlsize = 1, bestcol = "black", rangecol = rgb(0, 0,
  0, 0.3), slumpcol = grey(0.75), plotname = TRUE, ash = FALSE)
}
\arguments{
\item{allparameters}{Usual parameters from the Clam script. Check the manual.}

\item{saveData}{Boolean to keep Bacon outputs in the environment. Default: FALSE.}

\item{allparameters}{Usual parameters from the Clam script. Check the manual.}

\item{saveData}{Boolean to keep Bacon outputs in the environment. Default: FALSE.}
}
\description{
Run Clam.

Run Clam.
}
\examples{
Clam()
Clam()
}
\keyword{Age-depth}
\keyword{Chronology}
\keyword{model}

