% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Connect.R
\name{Connect}
\alias{Connect}
\title{Connection to the database}
\usage{
Connect(user, pass = -1, host = "localhost")
}
\arguments{
\item{user}{User.}

\item{pass}{Password. Default triggers the prompt.}

\item{host}{Adresse of the server. Default 'localhost'.}
}
\description{
This function allows you to connect to AFPALdb.
}
\examples{
con=Connect(user="root",pass="root")
}
\keyword{connect}

