% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Corplot.R
\name{correlationPlot}
\alias{correlationPlot}
\title{Generate a correlation plot.}
\usage{
correlationPlot(x1, x2, xmn, xmx, xstep, ymn = 0, ymx = xmx, ystep,
  Contour = FALSE, exportPDF = TRUE, savePDF = paste(getwd(),
  "/CorrelationPlot.pdf", sep = ""), title = "CorrelationPlot",
  xlab = "Age", ylab = "Width")
}
\arguments{
\item{x1}{Record to be compared. x1 must be a two-column matrix/data frame with time in the first column /and the variable to compare.}

\item{x2}{Idem.}

\item{xmn}{Date of beginning of the comparison.}

\item{xmx}{Date of end of the comparison.}

\item{xstep}{Temporal interval between two measures.}

\item{ymn}{Minimum size of window to be measured.}

\item{ymx}{Maximum size of window to be measured. Default is a global measure of the correlation of the two records.}

\item{ystep}{Steps of the correlation window.}

\item{Contour}{Boolean. If TRUE, contours at every 0.1 point of correlation will be added to the figure.}

\item{exportPDF}{Boolean. If TRUE, the graphic will be saved in the file named 'savePDF'.}

\item{savePDF}{Address and Name of the file. Default if 'CorrelationPlot.pdf' in the working directory.}

\item{title}{Title to display on the graphic.}

\item{xlab}{Label of the x-axis.}

\item{ylab}{Label of the y-axis.}
}
\description{
This figure measures the correlation between two records at different periods and with different time windows.
}
\examples{
d1=getData(Citation_Key="Chase_etal_2013",Record_Name="SWP-1-1",Proxy.Uncer=FALSE,Chrono.Uncer=FALSE,Proxy="d15N") ; d1=cbind(d1$Chronology,d1$Data$Data)
d2=getData(Citation_Key="Chase_etal_2013",Record_Name="SWP-1-1",Proxy.Uncer=FALSE,Chrono.Uncer=FALSE,Proxy="d13C") ; d2=cbind(d2$Chronology,d2$Data$Data)
correlationPlot(d1,d2,3000,8000,100,ymn=0,ystep=100,Contour=TRUE)
}
\keyword{Plot}

