% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MonteCarloStacking.R
\name{exportMonteCarloInterpolation}
\alias{exportMonteCarloInterpolation}
\title{Export interpolated reconstructions into .csv and .pdf files.}
\usage{
exportMonteCarloInterpolation(dat, saveData = getwd())
}
\arguments{
\item{dat}{The classic Monte-Carlo object generated by Interpolation.init().}

\item{saveData}{Folder where to save the outputs (in the working folder by default).}
}
\description{
Export interpolated reconstructions into .csv and .pdf files.
}
\examples{
exportMonteCarloInterpolation(Pella.init)
}
\keyword{Monte-Carlo}
\keyword{Plot}

