% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExtractTRACE.R
\name{extractTRACE}
\alias{extractTRACE}
\title{Extract data from TRACE.}
\usage{
extractTRACE(gridcells, var, month = 1, decadal = TRUE, layer = "",
  exportCSV = FALSE, saveCSV = paste(getwd(), "/Extract_", var, ".csv", sep
  = ""), tmin = -50, tmax = 22000)
}
\arguments{
\item{gridcells}{Coordinates of the pixels of interest (lon1, lat1, lon2, lat2, ...).}

\item{var}{Variable of interest (TS/PRECC/PRECT/AI/MAP/MAT/PET/TSMN/TSMX/SOLIN).}

\item{month}{A vector of values comprised between 1 and 12.}

\item{decadal}{Boolean to select between annual (FALSE) or averaged by decade (TRUE, default) timeseries.}

\item{layer}{The layer to extract from the netcdf file.}

\item{exportCSV}{Boolean to export the results in a csv.}

\item{saveCSV}{Address of the file. Default is "Extract_var.csv" in the working directory.}

\item{tmin}{Subset of the 22,000 years of simulation.}

\item{tmax}{Subset of the 22,000 years of simulation.}
}
\description{
Extract data from TRACE.
}
\examples{
TS=extractTRACE(c(0,0,5,5),"TS",month=4:6,exportCSV=TRUE)
}
\keyword{TRaCE}

