% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Connect.R
\name{forceConnect}
\alias{forceConnect}
\title{Connection to the database}
\usage{
forceConnect(user = -1, pass = -1, host = -1)
}
\arguments{
\item{user}{User.}

\item{pass}{Password. Default triggers the prompt.}

\item{host}{Adresse of the server. Default 'localhost'.}
}
\description{
This function forces a connection to AFPALdb. Called by any function if the user didn't use the Connect() function.
}
\examples{
con=forceConnect()
}
\keyword{forceConnect}

