% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_password.R
\name{get_password}
\alias{get_password}
\title{Get a password without echoing}
\usage{
get_password()
}
\description{
This function allows you to get a password without echoing.
}
\examples{
pass=get_password()
}
\keyword{password}

