% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getTable.R
\name{getAgeList}
\alias{getAgeList}
\title{Get the Age table.}
\usage{
getAgeList(ORDERBY = "LabCode")
}
\arguments{
\item{ORDERBY}{Set of variables (separated by a comma) to order the results by. Results sorted by 'LabCode' by default.}
}
\description{
This function allows you to get all the entries from the AFPALdb.C14 table.
}
\examples{
Age=getAgeList()
Age=getAgeList("LabCode")
}
\keyword{getTable}

