% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getTable.R
\name{getAltitude}
\alias{getAltitude}
\title{Get the Altitude table.}
\usage{
getAltitude()
}
\description{
This function allows you to get all the entries from the AFPALdb.Altitude table.
}
\examples{
altitude=getAltitude()
}
\keyword{getTable}

