% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getTable.R
\name{getBio}
\alias{getBio}
\title{Get the bio table.}
\usage{
getBio()
}
\description{
This function allows you to get all the entries from the AFPALdb.bio table.
}
\examples{
bio=getBio()
}
\keyword{getTable}

