% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getTable.R
\name{getBiome}
\alias{getBiome}
\title{Get the Biome table.}
\usage{
getBiome()
}
\description{
This function allows you to get all the entries from the AFPALdb.Biome table.
}
\examples{
biome=getBiome()
}
\keyword{getTable}

