% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getTable.R
\name{getC14List}
\alias{getC14List}
\title{Get the C14 table.}
\usage{
getC14List()
}
\description{
This function allows you to get all the entries from the AFPALdb.C14 table.
}
\examples{
C14=getC14List()
}
\keyword{getC14List}

