% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getTable.R
\name{getChronoAgeList}
\alias{getChronoAgeList}
\title{Get the ChronoC14 table.}
\usage{
getChronoAgeList(ORDERBY = "Chrono_ID,LabCode")
}
\arguments{
\item{ORDERBY}{Set of variables (separated by a comma) to order the results by. Results sorted by 'Chrono_ID x LabCode' by default.}
}
\description{
This function allows you to get all the entries from the AFPALdb.ChronoC14 table.
}
\examples{
chronocAge=getChronoAgeList()
chronocAge=getChronoAgeList("LabCode")
}
\keyword{getTable}

