% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getSites.R
\name{getChronoC14List}
\alias{getChronoC14List}
\title{Get the ChronoC14 table.}
\usage{
getChronoC14List()
}
\description{
This function allows you to get all the entries from the AFPALdb.ChronoC14 table.
}
\examples{
chronodata=getChronoC14List()
}
\keyword{getChronoC14List}

