% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getSites.R
\name{getChronoDataList}
\alias{getChronoDataList}
\title{Get the ChronoData table.}
\usage{
getChronoDataList()
}
\description{
This function allows you to get all the entries from the AFPALdb.ChronoData table.
}
\examples{
chronodata=getChronoDataList()
}
\keyword{getChronoDataList}

