% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ColorGradient.R
\name{getColorGradient}
\alias{getColorGradient}
\title{Open a SSH tunnel with the server to open a distant file.}
\usage{
getColorGradient(name = "", reverse = FALSE)
}
\arguments{
\item{name}{Name of the gradient. If empty or mispelled, the user will have to select from the list of available gradients.}

\item{reverse}{Boolean to reverse the color gradient.}
}
\description{
This functions opens a SSH tunnel with the server to open a distant file.
}
\examples{
getColorGradient()
}
\keyword{Plot}

