% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getChronology.R, R/getData.R
\name{getData}
\alias{getData}
\title{Extract chronologies from the database}
\usage{
getData(Site_Name = "", Record_Name = "", Citation_Key = "", Proxy = "",
  Proxy.Uncer = FALSE, Dataset_ID = -1, Chronology = TRUE,
  Chrono.Uncer = FALSE)

getData(Site_Name = "", Record_Name = "", Citation_Key = "", Proxy = "",
  Proxy.Uncer = FALSE, Dataset_ID = -1, Chronology = TRUE,
  Chrono.Uncer = FALSE)
}
\arguments{
\item{Site_Name}{Name of the site.}

\item{Record_Name}{Name of the specific record requested.}

\item{Citation_Key}{Reference associated to the dataset requested.}

\item{Proxy}{Type of proxy of dataset requested.}

\item{Proxy.Uncer}{Boolean to indicate if uncertainties should be downloaded when available.}

\item{Dataset_ID}{ID of the requested dataset.}

\item{Chronology}{Boolean to indicate if achronology should be downloaded (default is TRUE).}

\item{Chrono.Uncer}{Boolean to indicate if uncertainties should be downloaded when available.}

\item{Chrono_ID}{ID of the requested dataset.}

\item{Site_Name}{Name of the site.}

\item{Record_Name}{Name of the specific record requested.}

\item{Citation_Key}{Reference associated to the dataset requested.}

\item{Proxy}{Type of proxy of dataset requested.}
}
\description{
This function allows you to extract JSON objects (chronology containers) from AFPALdb.

This function allows you to extract JSON objects (data containers) from AFPALdb.
}
\examples{
d0=getData() # Returns NULL
d1=getData(Site_Name="Seweweekspoort")
d2=getData(Proxy="d13C")
d3=getData(Citation_Key="Chase_etal_2013")
d4=getData(Citation_Key="Chase_etal_2013",Record_Name="SWP-1-1")
d0=getData()
d1=getData(Site_Name="Seweweekspoort")
d2=getData(Proxy="d13C")
d3=getData(Citation_Key="Chase_etal_2013")
d4=getData(Citation_Key="Chase_etal_2013",Record_Name="SWP-1-1",Proxy.Uncer=TRUE,Chrono.Uncer=TRUE)
}
\keyword{Chronology}
\keyword{Dataset}
\keyword{getData}

