% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getDataset.R
\name{getDatasetFromReference}
\alias{getDatasetFromReference}
\title{Get a list of datasets associated to a given reference.}
\usage{
getDatasetFromReference(Citation_Key, fullRef = FALSE,
  ORDERBY = "Record_Name")
}
\arguments{
\item{Citation_Key}{Citation key.}

\item{fullRef}{Boolean. TRUE returns all the details of the site.}

\item{ORDERBY}{Field name to sort the output table. Default 'Record_Name'.}
}
\description{
This function allows you to get a list of datasets associated to a given reference. This function DOES NOT return data, just information about the dataset.
}
\examples{
getDatasetFromReference("Chase_etal_2013")
getDatasetFromReference("Tierney_etal_2008",ORDERBY="Proxy")
}
\keyword{getDataset}

