% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getDataset.R
\name{getDatasetFromSite}
\alias{getDatasetFromSite}
\title{Get a list of datasets associated to a given site.}
\usage{
getDatasetFromSite(Site_ID = -1, Site_Name = "", Record_Name = "",
  fullRef = FALSE, ORDERBY = "Record_Name")
}
\arguments{
\item{fullRef}{Boolean. TRUE returns all the details of the site.}

\item{ORDERBY}{Field name to sort the output table. Default 'Record_Name'.}

\item{Citation_Key}{Citation key.}
}
\description{
This function allows you to get a list of datasets associated to a given site. This function DOES NOT return data, just information about the dataset.
}
\examples{
getDatasetFromSite(Site_ID=1)
getDatasetFromSite(Site_Name="Seweweekspoort")
getDatasetFromSite(Record_Name="SWP-1-1")
}
\keyword{getDataset}

