% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getSites.R
\name{getDatasetList}
\alias{getDatasetList}
\title{Get the Dataset table.}
\usage{
getDatasetList()
}
\description{
This function allows you to get all the entries from the AFPALdb.Reference table, except the data.
}
\examples{
dataset=getDatasetList()
}
\keyword{getDatasetList}

