% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getDataSSH.R
\name{getDataSSH}
\alias{getDataSSH}
\title{Open a SSH tunnel with the server to open a file.}
\usage{
getDataSSH(file, header = TRUE, sep = ",")
}
\arguments{
\item{file}{Complete address of the file.}

\item{header}{Boolean (used in read.table).}

\item{sep}{Column separator (used in read.table).}
}
\description{
This functions opens a SSH tunnel with the server to open a file.
}
\examples{
head(getDataSSH("/home/emmanuel.chevalier/Bureau/*.csv"))
}
\keyword{connection}

