% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getTable.R
\name{getDistrib}
\alias{getDistrib}
\title{Get the distrib table.}
\usage{
getDistrib()
}
\description{
This function allows you to get all the entries from the AFPALdb.distrib table.
}
\examples{
distrib=getDistrib()
}
\keyword{getTable}

