% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getTable.R
\name{getPET}
\alias{getPET}
\title{Get the PET table.}
\usage{
getPET()
}
\description{
This function allows you to get all the entries from the AFPALdb.PET table.
}
\examples{
pet=getPET()
}
\keyword{getTable}

