% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getTable.R
\name{getPrecip}
\alias{getPrecip}
\title{Get the precip table.}
\usage{
getPrecip()
}
\description{
This function allows you to get all the entries from the AFPALdb.precip table.
}
\examples{
precip=getPrecip()
}
\keyword{getTable}

