% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getReference.R
\name{getReferenceFromDataID}
\alias{getReferenceFromDataID}
\title{Get the reference associated to a dataset.}
\usage{
getReferenceFromDataID(Dataset_ID, fullRef = FALSE)
}
\arguments{
\item{fullRef}{Boolean. TRUE returns all the details of the reference.}

\item{Region}{Name of the region.}
}
\description{
This function allows you to get the reference associated to a dataset.
}
\examples{
getReferenceFromDataID(1)
getReferenceFromDataID(2,TRUE)
}
\keyword{getReference}

