% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getReference.R
\name{getReferenceFromRegion}
\alias{getReferenceFromRegion}
\title{Get a list of studies performed in a given Region.}
\usage{
getReferenceFromRegion(Region, fullRef = FALSE, ORDERBY = "Citation_Key")
}
\arguments{
\item{Region}{Name of the region.}

\item{fullRef}{Boolean. TRUE returns all the details of the reference.}

\item{ORDERBY}{Field name to sort the output table. Default 'Citation_Key'.}
}
\description{
This function allows you to get a list of studies performed in a given Region.
}
\examples{
getReferenceFromRegion("SRZ")
getReferenceFromRegion("WRZ",TRUE,ORDERBY="Site_Name")
}
\keyword{getReference}

