% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getReference.R
\name{getReferenceFromSite}
\alias{getReferenceFromSite}
\title{Get a list of references associated to a given site.}
\usage{
getReferenceFromSite(Site_ID = -1, Site_Name = "", Record_Name = "",
  fullRef = FALSE, ORDERBY = "Citation_Key")
}
\arguments{
\item{Site_ID}{ID of the site.}

\item{Site_Name}{Name of the site.}

\item{Record_Name}{Name of the record.}

\item{fullRef}{Boolean. TRUE returns all the details of the reference.}

\item{ORDERBY}{Field name to sort the output table. Default 'Citation_Key'.}
}
\description{
This function allows you to get a list of references associated to a given site.
}
\examples{
d1=getReferenceFromSite(Site_Name="Seweweekspoort")
d2=getReferenceFromSite(Site_ID=1,fullRef=FALSE,ORDERBY="Journal")
d3=getReferenceFromSite(Record_Name="SWP-1-1",fullRef=TRUE)
}
\keyword{getReference}

