% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getReference.R
\name{getReferenceFromXY}
\alias{getReferenceFromXY}
\title{Get a list of studies performed in a given area.}
\usage{
getReferenceFromXY(xmn, xmx, ymn, ymx, fullRef = FALSE,
  ORDERBY = "Citation_Key")
}
\arguments{
\item{xmn}{Lower bound of the longitudinal range.}

\item{xmx}{Upper bound of the longitudinal range.}

\item{ymn}{Lower bound of the latitudinal range.}

\item{ymx}{Upper bound of the latitudinal range.}

\item{fullRef}{Boolean. TRUE returns all the details of the reference.}

\item{ORDERBY}{Field name to sort the output table. Default 'Citation_Key'.}
}
\description{
This function allows you to get a list of studies performed in a given area.
}
\examples{
getReferenceFromXY(20,30,-30,-20,ORDERBY="Site_Name")
getReferenceFromXY(25,30,-30,-20,fullRef=TRUE)
getReferenceFromXY(25,20,-30,-20)
}
\keyword{getReference}

