% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getTable.R
\name{getReferenceList}
\alias{getReferenceList}
\title{Get the Reference table.}
\usage{
getReferenceList(ORDERBY = "Citation_Key")
}
\arguments{
\item{ORDERBY}{Set of variables (separated by a comma) to order the results by. Results sorted by 'Citation_Key' by default.}
}
\description{
This function allows you to get all the entries from the AFPALdb.Reference table.
}
\examples{
ref=getReferenceList()
ref=getReferenceList("Citation_Key")
}
\keyword{getTable}

