% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getSite.R
\name{getSiteFromDataID}
\alias{getSiteFromDataID}
\title{Get the site associated to a dataset.}
\usage{
getSiteFromDataID(Dataset_ID, fullRef = FALSE)
}
\arguments{
\item{fullRef}{Boolean. TRUE returns all the details of the site.}

\item{Region}{Name of the region.}
}
\description{
This function allows you to get the site associated to a dataset.
}
\examples{
getSiteFromDataID(1)
getSiteFromDataID(2,TRUE)
}
\keyword{getSite}

