% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getSite.R
\name{getSiteFromReference}
\alias{getSiteFromReference}
\title{Get a list of references associated to a given site.}
\usage{
getSiteFromReference(Citation_Key, fullRef = FALSE, ORDERBY = "Site_Name")
}
\arguments{
\item{Citation_Key}{Citation key.}

\item{fullRef}{Boolean. TRUE returns all the details of the site.}

\item{ORDERBY}{Field name to sort the output table. Default 'Site_Name'.}
}
\description{
This function allows you to get a list of references associated to a given site.
}
\examples{
getSiteFromReference("Chase_etal_2013")
getSiteFromReference("Tierney_etal_2008",TRUE)
}
\keyword{getSite}

