% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getSite.R
\name{getSiteFromRegion}
\alias{getSiteFromRegion}
\title{Get a list of sites part of a given Region.}
\usage{
getSiteFromRegion(Region, fullRef = FALSE, ORDERBY = "Site_Name")
}
\arguments{
\item{Region}{Name of the region.}

\item{fullRef}{Boolean. TRUE returns all the details of the reference.}

\item{ORDERBY}{Field name to sort the output table. Default 'Site_Name'.}
}
\description{
This function allows you to get a list of sites part of a given Region.
}
\examples{
getSiteFromRegion("SRZ")
getSiteFromRegion("WRZ",TRUE)
}
\keyword{getSite}

