% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getTable.R
\name{getTaxalist}
\alias{getTaxalist}
\title{Get the Taxalist table.}
\usage{
getTaxalist()
}
\description{
This function allows you to get all the entries from the AFPALdb.Taxalist table.
}
\examples{
taxalist=getTaxalist()
}
\keyword{getTable}

