% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getTable.R
\name{getTmpr_max}
\alias{getTmpr_max}
\title{Get the tmpr_max table.}
\usage{
getTmpr_max()
}
\description{
This function allows you to get all the entries from the AFPALdb.tmpr_max table.
}
\examples{
tmpr_max=getTmpr_max()
}
\keyword{getTable}

