% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getTable.R
\name{getTmpr_mean}
\alias{getTmpr_mean}
\title{Get the tmpr_mean table.}
\usage{
getTmpr_mean()
}
\description{
This function allows you to get all the entries from the AFPALdb.tmpr_mean table.
}
\examples{
tmpr_mean=getTmpr_mean()
}
\keyword{getTable}

