% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getTable.R
\name{getTmpr_min}
\alias{getTmpr_min}
\title{Get the tmpr_min table.}
\usage{
getTmpr_min()
}
\description{
This function allows you to get all the entries from the AFPALdb.tmpr_min table.
}
\examples{
tmpr_min=getTmpr_min()
}
\keyword{getTable}

