% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MonteCarloStacking.R
\name{Interpolation.init}
\alias{Interpolation.init}
\title{Function used to initiate the Monte-Carlo process. It will extract the data from FPALdb and group them into a Monte-Carlo object.}
\usage{
Interpolation.init(sites, variable, time.range, time.resol, NREP = 50000,
  Citation_Keys = rep("", length(sites)))
}
\arguments{
\item{sites}{List of sites to be used in the process.}

\item{variable}{Variable of interest.}

\item{time.range}{Minimum and maximum values of th period of interest. Does not necessary encompass the total temporal range of the records.}

\item{time.resol}{Resolution of the record.}

\item{NREP}{Number of scenarios to generate (default 50,000).}

\item{Citation_Keys}{Number of scenarios to generate (default 50,000).}
}
\description{
Function used to initiate the Monte-Carlo process. It will extract the data from FPALdb and group them into a Monte-Carlo object.
}
\examples{
Pella.init=Interpolation.init(sites=c("PEL-1-1","PEL-1-4a"),variable="ReconAI",time.range=c(0,50000),time.resol=100,NREP=1000,Citation_Keys=c("Lim_etal_2016","Lim_etal_2016"))
}
\keyword{Monte-Carlo}

