% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plot.R
\name{makeTransparent}
\alias{makeTransparent}
\title{Add transparency to a color.}
\usage{
makeTransparent(color, alpha = 0.5)
}
\arguments{
\item{color}{List of colors.}

\item{alpha}{Degree of transparency (0: full transparency, 1: no transparency).}
}
\description{
Add transparency to a color.
}
\examples{
plot(1:20,1:20,col=getColorGradient("Rd2Bl20"),pch=15,cex=5)
points(1:20,1:20+2,col=makeTransparent(getColorGradient("Rd2Bl20"),alpha=0.5),pch=15,cex=5)

}
\keyword{Plot}

