% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MonteCarloStacking.R
\name{MonteCarlo}
\alias{MonteCarlo}
\title{Function to run all the Monte-Carlo interpolation and stacking process at once.}
\usage{
MonteCarlo(sites, variable, time.range, time.resol, NREP = 50000,
  Citation_Keys = rep("", length(sites)), remove = NA, ref = "",
  overlap = NA, exportData = FALSE, saveData = getwd())
}
\arguments{
\item{sites}{List of sites to be used in the process.}

\item{variable}{Variable of interest.}

\item{time.range}{Minimum and maximum values of th period of interest. Does not necessary encompass the total temporal range of the records.}

\item{time.resol}{Resolution of the record.}

\item{NREP}{Number of scenarios to generate (default 50,000).}

\item{Citation_Keys}{Number of scenarios to generate (default 50,000).}

\item{remove}{A vector that contains the values necessary to run the 'removeSample()' function. Should be formatted as follow (n1,site1, n2,site2, n3,site3).}

\item{ref}{String that indicates the sequence of reference. (Used for centring)}

\item{overlap}{Vector that indicates for each site the period to consider for the overlap, e.g. c("site1",0,10000,"site2",3000,11000,"site3",350,8000). (Used for centring)}

\item{exportData}{Boolean (default FALSE) to save the individual interpolations.}

\item{saveData}{Folder where to save the outputs (in the working folder by default).}
}
\description{
Function to run all the Monte-Carlo interpolation and stacking process at once.
}
\examples{
Pella = MonteCarlo(sites=c("PEL-1-1","PEL-1-4a"),variable="ReconAI",time.range=c(0,50000),time.resol=100,NREP=1000,Citation_Keys=c("Lim_etal_2016","Lim_etal_2016"),remove=c(21,"PEL-1-1"),ref="PEL-1-1",overlap=c("PEL-1-4a",100,1200),exportData=TRUE)
}
\keyword{Monte-Carlo}

