% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MonteCarloStacking.R
\name{MonteCarloCentring}
\alias{MonteCarloCentring}
\title{Automatically calculates an optimum centring value for the stacking.}
\usage{
MonteCarloCentring(dat, ref, overlap)
}
\arguments{
\item{dat}{The classic Monte-Carlo object generated by Interpolation.init().}

\item{ref}{String that indicates the sequence of reference.}

\item{overlap}{Vector that indicates for each site the period to consider for the overlap, e.g. c("site1",0,10000,"site2",3000,11000,"site3",350,8000)}
}
\description{
Automatically calculates an optimum centring value for the stacking.
}
\examples{
Pella.init=MonteCarloCentring(Pella.init,"PEL-1-1",c("PEL-1-4a",100,1200))
}
\keyword{Monte-Carlo}

