% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MonteCarloStacking.R
\name{MonteCarloInterpolation}
\alias{MonteCarloInterpolation}
\title{Interpolate individually each reconstruction.}
\usage{
MonteCarloInterpolation(dat, exportData = FALSE, saveData = getwd())
}
\arguments{
\item{dat}{The classic Monte-Carlo object generated by Interpolation.init().}

\item{exportData}{Boolean (default FALSE) to save the individual interpolations.}

\item{saveData}{Folder where to save the outputs (in the working folder by default).}
}
\description{
Interpolate individually each reconstruction.
}
\examples{
Pella.init=MonteCarloInterpolation(Pella.init,exportData=TRUE)
}
\keyword{Monte-Carlo}

