% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/openPDF.R
\name{openArticle}
\alias{openArticle}
\title{Function that opens pdf files from a reference or an author.}
\usage{
openArticle(Citation_Key = "", Author = "")
}
\arguments{
\item{Citation_Key}{Key of the reference.}

\item{Author}{Name of the author. Selection on the terminal possible if many references correspond to the name.}
}
\description{
Function that opens pdf files from a reference or an author.
}
\examples{
openArticle("Baker_etal_2014")
openArticle(Author="Scott")
}
\keyword{pdf}

