% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TRACEfunctions.R
\name{openTRACE}
\alias{openTRACE}
\title{Open a TRACE file}
\usage{
openTRACE(var, month = 1, decadal = TRUE, layer = "")
}
\arguments{
\item{var}{Variable of interest (TS/PRECC/PRECL/AI/MAP/MAT/PET/TSMN/TSMX/SOLIN).}

\item{month}{A value between 1 and 12.}

\item{decadal}{Boolean to select between annual (FALSE) or averaged by decade (TRUE, default) timeseries.}

\item{layer}{The layer to extract from the netcdf file.}
}
\description{
Open a TRACE file.
}
\examples{
TS=openTRACE("TS",month=4,decadal=TRUE)
}
\keyword{TRaCE}

