% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plot.R
\name{plotCountries}
\alias{plotCountries}
\title{Add countries to a plot.}
\usage{
plotCountries(EXT = c(-180, 180, -90, 90), add = FALSE)
}
\arguments{
\item{EXT}{Extension of the plot( xmin, xmax, ymin, ymax).}

\item{add}{Boolean. If TRUE (default), add countries borders to an existing plot.}
}
\description{
Add countries to a plot.
}
\examples{
plotCountries(EXT=c(0,40,-40,0),add=FALSE)
}
\keyword{Plot}

