% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plot.R
\name{plotCountriesByName}
\alias{plotCountriesByName}
\title{Add countries to a plot.}
\usage{
plotCountriesByName(countries = c("South Africa"), add = FALSE)
}
\arguments{
\item{countries}{List of countries to be plotted.}

\item{add}{Boolean. If TRUE (default), add countries borders to an existing plot.}
}
\description{
Add countries to a plot.
}
\examples{
plotCountriesByName(countries=c("South Africa","Lesotho","Namibia","Botswana","Swaziland"),add=FALSE)
}
\keyword{Plot}

