% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotTRACE.R
\name{plotTRACE}
\alias{plotTRACE}
\title{Extract data from TRACE.}
\usage{
plotTRACE(dat, year = 1, zlim = range(dat[, , year]), exportPDF = FALSE,
  savePDF = paste(getwd(), "/plotTRACE.pdf", sep = ""), add.contours = TRUE,
  add.countries = TRUE, WIDTH = 5.51, HEIGHT = 5.51, legend = "bottom",
  EXT = c(-180, 180, -90, 90), COL = rev(getColorGradient("Rd2Bl20")))
}
\arguments{
\item{dat}{Matrix to be plotted.}

\item{year}{Year to plot. (1=22000 cal BP)}

\item{zlim}{Range of values.}

\item{exportPDF}{Boolean. TRUE to generate a pdf of the figure.}

\item{savePDF}{Address and name of the pdf. Default is "plotTRACE.pdf" in the working directory.}

\item{add.contours}{Boolean to add contours on the figure.}

\item{WIDTH}{The width of the figure in inches.}

\item{HEIGHT}{The height of the figure in inches.}

\item{legend}{Position of the legend (top/right/bottom/left/none).}

\item{EXT}{Extension of the plot( xmin, xmax, ymin, ymax).}

\item{add.contries}{Boolean to add country borders on the figure.}
}
\description{
Extract data from TRACE.
}
\examples{
plotTRACE(dat=openTRACE("TS",month=1,decadal=TRUE),year=1,exportPDF=TRUE,EXT=c(-180,180,-90,90),legend='bottom')
plotTRACE(dat=openTRACE("PRECC",month=8,decadal=TRUE),year=1,exportPDF=TRUE,EXT=c(-11.25,56.2,-37.5,41.2),legend='right')
}
\keyword{Plot}
\keyword{TRaCE}

