% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RAFPALdb.R
\name{test_fct}
\alias{test_fct}
\title{A Cat Function}
\usage{
test_fct()
}
\arguments{
\item{love}{Do you love cats? Defaults to TRUE.}
}
\description{
This function allows you to express your love of cats.
}
\examples{
cat_function()
}
\keyword{cats}

