% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Connect.R
\name{Connect}
\alias{Connect}
\title{Connection to the database}
\usage{
Connect(user = -1, pass = -1, host = -1)
}
\arguments{
\item{user}{User.}

\item{pass}{Password.}

\item{host}{Adresse of the server. Default 'localhost'.}
}
\description{
This function allows you to connect to AFPALdb.
}
\examples{
Connect(user="root",pass="root123")
}
\keyword{connection}

