% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Connect.R
\name{Connect}
\alias{Connect}
\title{Connection to the database}
\usage{
Connect(user, pass)
}
\arguments{
\item{user}{User.}

\item{pass}{Password.}
}
\description{
This function allows you to connect to AFPALdb.
}
\examples{
con=Connect(user="root",pass="root")
}
\keyword{connect}

