% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getSites.R
\name{getChronologyList}
\alias{getChronologyList}
\title{Get the Chronology table.}
\usage{
getChronologyList()
}
\description{
This function allows you to get all the entries from the AFPALdb.Chronology table, except the data.
}
\examples{
dataset=getChronologyList()
}
\keyword{getChronologyList}

