% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getSites.R
\name{getReferenceList}
\alias{getReferenceList}
\title{Get the Reference table.}
\usage{
getReferenceList()
}
\description{
This function allows you to get all the entries from the AFPALdb.Reference table.
}
\examples{
ref=getReferenceList()
}
\keyword{getReferenceList}

