% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getTable.R
\name{getRefSiteList}
\alias{getRefSiteList}
\title{Get the RefSite table.}
\usage{
getRefSiteList()
}
\description{
This function allows you to get all the entries from the AFPALdb.RefSite table.
}
\examples{
refsite=getRefSiteList()
}
\keyword{getRefSiteList}

