% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getSites.R
\name{getSiteList}
\alias{getSiteList}
\title{Get the Site table.}
\usage{
getSiteList()
}
\description{
This function allows you to get all the entries from the AFPALdb.Site table.
}
\examples{
sites=getSiteList()
}
\keyword{getSiteList}

