% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getTable.R
\name{getSiteRegionList}
\alias{getSiteRegionList}
\title{Get the SiteRegion table.}
\usage{
getSiteRegionList()
}
\description{
This function allows you to get all the entries from the AFPALdb.SiteRegion table.
}
\examples{
siteregion=getSiteRegionList()
}
\keyword{getSiteRegionList}

